Moved Face into IGMdataitem namespace
