+++ Core/Core.csproj
@@ -179,6 +179,7 @@
     <Compile Include="Menu\IGMDataItem\Base.cs" />
     <Compile Include="Menu\IGMDataItem\Box.cs" />
     <Compile Include="Menu\IGMDataItem\Empty.cs" />
+    <Compile Include="Menu\IGMDataItem\Face.cs" />
     <Compile Include="Menu\IGMDataItem\I_Color.cs" />
     <Compile Include="Menu\IGMDataItem\I_Data.cs" />
     <Compile Include="Menu\IGMDataItem\I_FontColor.cs" />
@@ -773,7 +774,6 @@
     <Compile Include="Menu\Images\Icons.cs" />
     <Compile Include="Menu\Images\Icons.ID.cs" />
     <Compile Include="Menu\Images\Icons.InsertCustomEntries.cs" />
-    <Compile Include="Menu\IGMDataItem\IGMDataItem_Face.cs" />
     <Compile Include="Menu\IGMDataItem\IGMDataItem_Icon.cs" />
     <Compile Include="Menu\IGMDataItem\IGMDataItem_Int.cs" />
     <Compile Include="Menu\IGMDataItem\IGMDataItem_String.cs" />
